+++ src/scripts/magic-item-entry/MagicItemTable.js
@@ -1,4 +1,5 @@
 import { MAGICITEMS } from "../config";
+import Logger from "../lib/Logger";
 import { MagicItemHelpers } from "../magic-item-helpers";
 import { AbstractMagicItemEntry } from "./AbstractMagicItemEntry";
 
@@ -18,18 +19,21 @@ export class MagicItemTable extends AbstractMagicItemEntry {
       const collectionId = result.results[0].documentCollection;
       const id = result.results[0].documentId;
       const pack = game.collections.get(collectionId) || game.packs.get(collectionId);
-
-      const entity = pack.getDocument ? await pack.getDocument(id) : pack.get(id);
-      if (entity) {
-        let item = (await actor.createEmbeddedDocuments("Item", [entity]))[0];
-        const chatData = await item.use({}, { createMessage: false });
-        // Fix https://github.com/PwQt/magic-items-2/issues/22
-        if (!game.modules.get("ready-set-roll-5e")?.active) {
-          ChatMessage.create(
-            mergeObject(chatData, {
-              "flags.dnd5e.itemData": item,
-            })
-          );
+      if (!pack) {
+        Logger.warn(`Cannot retrieve pack for if ${collectionId}`, true);
+      } else {
+        const entity = pack.getDocument ? await pack.getDocument(id) : pack.get(id);
+        if (entity) {
+          let item = (await actor.createEmbeddedDocuments("Item", [entity]))[0];
+          const chatData = await item.use({}, { createMessage: false });
+          // Fix https://github.com/PwQt/magic-items-2/issues/22
+          if (!game.modules.get("ready-set-roll-5e")?.active) {
+            ChatMessage.create(
+              mergeObject(chatData, {
+                "flags.dnd5e.itemData": item,
+              }),
+            );
+          }
         }
       }
     }