+++ src/scripts/magic-item/OwnedMagicItem.js
@@ -1,6 +1,6 @@
 import { MAGICITEMS } from "../config";
 import CONSTANTS from "../constants/constants";
-import { warn } from "../lib/lib";
+import Logger from "../lib/Logger";
 import { MagicItemHelpers } from "../magic-item-helpers";
 import { OwnedMagicItemFeat } from "../magic-item-owned-entry/OwnedMagicItemFeat";
 import { OwnedMagicItemSpell } from "../magic-item-owned-entry/OwnedMagicItemSpell";
@@ -8,8 +8,8 @@ import { OwnedMagicItemTable } from "../magic-item-owned-entry/OwnedMagicItemTab
 import { MagicItem } from "./MagicItem";
 
 export class OwnedMagicItem extends MagicItem {
-  constructor(item, actor, magicItemActor) {
-    super(item.flags.magicitems);
+  constructor(item, actor, magicItemActor, flagsData) {
+    super(flagsData);
     this.uuid = item.uuid;
     this.id = item.id;
     this.item = item;
@@ -17,7 +17,7 @@ export class OwnedMagicItem extends MagicItem {
     this.name = item.name;
     this.img = item.img;
     this.pack = item.pack;
-    this.uses = parseInt("uses" in item.flags.magicitems ? item.flags.magicitems.uses : this.charges);
+    this.uses = parseInt("uses" in flagsData ? flagsData.uses : this.charges);
 
     this.rechargeableLabel = this.rechargeable
       ? `(${game.i18n.localize("MAGICITEMS.SheetRecharge")}: ${this.rechargeText} ${
@@ -89,7 +89,7 @@ export class OwnedMagicItem extends MagicItem {
   rollByName(itemName) {
     let found = this.ownedEntries.filter((entry) => entry.name === itemName);
     if (!found.length) {
-      warn(game.i18n.localize("MAGICITEMS.WarnNoMagicItemSpell") + itemName, true);
+      Logger.warn(game.i18n.localize("MAGICITEMS.WarnNoMagicItemSpell") + itemName, true);
       return;
     }
     found[0].roll();
@@ -233,7 +233,7 @@ export class OwnedMagicItem extends MagicItem {
     this.item
       .update({
         flags: {
-          magicitems: this.serializeData(),
+          [CONSTANTS.MODULE_ID]: this.serializeData(),
         },
       })
       .then(() => {