+++ src/scripts/magicitemsheet.js
@@ -60,35 +60,35 @@ export class MagicItemSheet {
       if (this.actor.hasItemsFeats()) {
         await this.renderTemplate(
           "magic-item-feat-sheet.html",
-          "magic-items-feats-content",
+          "magic-items-2-feats-content",
           "features",
-          "inventory-list"
+          "inventory-list",
         );
       }
       if (this.actor.hasItemsSpells()) {
         await this.renderTemplate(
           "magic-item-spell-sheet.html",
-          "magic-items-spells-content",
+          "magic-items-2-spells-content",
           "spellbook",
-          "inventory-list"
+          "inventory-list",
         );
       }
     } else {
       if (this.actor.hasItemsFeats()) {
         await this.renderTemplate(
           "magic-item-feat-sheet-v2.hbs",
-          "magic-items-feats-content",
+          "magic-items-2-feats-content",
           "features",
-          "features-list"
+          "features-list",
         );
         this.html.find(".item-tooltip").each((idx, el) => this.addToolTips(el));
       }
       if (this.actor.hasItemsSpells()) {
         await this.renderTemplate(
           "magic-item-spell-sheet-v2.hbs",
-          "magic-items-spells-content",
+          "magic-items-2-spells-content",
           "spells",
-          "spells-list"
+          "spells-list",
         );
         this.html.find(".item-tooltip").each((idx, el) => this.addToolTips(el));
       }
@@ -132,9 +132,10 @@ export class MagicItemSheet {
    */
   addToolTips(element) {
     if ("tooltip" in element.dataset) return;
-    const target = element.closest("[data-item-id], [data-uuid]");
+    const target = element.closest("[data-item-id], [data-uuid]"); // TODO why data-uuid ?
     const uuid = target.dataset?.itemUuid;
     if (!uuid) return;
+    // TODO why data-uuid ?
     element.dataset.tooltip = `
       <section class="loading" data-uuid="${uuid}"><i class="fas fa-spinner fa-spin-pulse"></i></section>
     `;
@@ -164,12 +165,12 @@ export class MagicItemSheet {
 
   static handleActorItemUsesChangeEvents(html, actor) {
     actor.items.forEach((item) => {
-      html.find(`input[data-item-id="magicitems.${item.id}.uses"]`).change((evt) => {
+      html.find(`input[data-item-uses="magic-items-2.${item.id}.uses"]`).change((evt) => {
         item.setUses(MagicItemHelpers.numeric(evt.currentTarget.value, item.uses));
         item.update();
       });
       item.ownedEntries.forEach((entry) => {
-        html.find(`input[data-item-id="magicitems.${item.id}.${entry.id}.uses"]`).change((evt) => {
+        html.find(`input[data-item-uses="magic-items-2.${item.id}.${entry.id}.uses"]`).change((evt) => {
           entry.uses = MagicItemHelpers.numeric(evt.currentTarget.value, entry.uses);
           item.update();
         });